move effect funcs to webcam_api
