+++ tools/webcam/webcam_apis/utils/__init__.py
@@ -2,9 +2,11 @@
 from .buffer import BufferManager
 from .event import EventManager
 from .message import FrameMessage, Message, VideoEndingMessage
-from .misc import limit_max_fps, load_image_from_disk_or_url
+from .misc import (copy_and_paste, expand_and_clamp, limit_max_fps,
+                   load_image_from_disk_or_url, screen_matting)
 
 __all__ = [
     'BufferManager', 'EventManager', 'FrameMessage', 'Message',
-    'limit_max_fps', 'VideoEndingMessage', 'load_image_from_disk_or_url'
+    'limit_max_fps', 'VideoEndingMessage', 'load_image_from_disk_or_url',
+    'screen_matting', 'expand_and_clamp', 'copy_and_paste'
 ]
